fixed cache problem
Say there was a function named myFunc. If you called myFunc with one set of parameters and then immediately called myFunc again with a new set of parameters (without calling any argParser functions between the two myFunc calls), the cached copy of the args from the first myFunc call was being used for the second myFunc call. This is now fixed.